addons/voice/requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0

# Audio processing
soundfile>=0.12.0
librosa>=0.10.0
demucs>=4.0.0

# Voice cloning
TTS>=0.22.0  # Coqui TTS with XTTS

# Video/Audio download
yt-dlp>=2024.0.0

# Web search
tavily-python>=0.3.0

# LLM integration
llama-cpp-python>=0.2.0

# Embeddings
fastembed>=0.2.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0.0  # Pretty console output
click>=8.0.0  # CLI
pydantic>=2.0.0  # Data validation
jinja2>=3.0.0  # Templates

# Optional: GPU acceleration
# torch with CUDA - install separately based on your GPU
